refactor(tests): split unit_tests.cpp into per-fixture test files
Split the 449-line tests/unit_tests.cpp into 6 per-fixture files
(test_database.cpp, test_orm.cpp, test_monitoring.cpp,
test_security.cpp, test_async.cpp, test_query_builder.cpp).
Update tests/CMakeLists.txt to compile all sources into the same
database_unit_tests executable so gtest_discover_tests continues to
discover the existing tests unchanged.

Per-fixture editing now triggers only that fixture's recompile, while
the test count and CTest output remain identical.

Part of #577. Closes #580.
