--- tests/unit_tests.cpp
@@ -1,450 +1,0 @@
-// BSD 3-Clause License
-// Copyright (c) 2025, 🍀☀🌕🌥 🌊
-// See the LICENSE file in the project root for full license information.
-
-#include <chrono>
-#include <future>
-#include <gtest/gtest.h>
-#include <memory>
-#include <thread>
-
-#include <kcenon/database/async/async_operations.h>
-#include <kcenon/database/core/database_context.h>
-#include <kcenon/database/database_manager.h>
-#include <kcenon/database/database_types.h>
-#include <kcenon/database/monitoring/performance_monitor.h>
-#include <kcenon/database/orm/entity.h>
-#include <kcenon/database/security/secure_connection.h>
-
-using namespace database;
-using namespace database::orm;
-using namespace database::monitoring;
-using namespace database::security;
-using namespace database::async;
-
-// Test fixture for database tests
-class DatabaseTest : public ::testing::Test {
-protected:
-  std::shared_ptr<database_context> context_;
-  std::shared_ptr<database_manager> db_mgr_;
-
-  void SetUp() override {
-    // Test setup with dependency injection
-    context_ = std::make_shared<database_context>();
-    db_mgr_ = std::make_shared<database_manager>(context_);
-  }
-
-  void TearDown() override {
-    // Cleanup
-    if (db_mgr_) {
-      db_mgr_->disconnect_result();
-    }
-  }
-};
-
-// Basic database manager tests
-TEST_F(DatabaseTest, DatabaseManagerDependencyInjection) {
-  // Test that dependency injection pattern works correctly
-  auto context1 = std::make_shared<database_context>();
-  auto db1 = std::make_shared<database_manager>(context1);
-
-  auto context2 = std::make_shared<database_context>();
-  auto db2 = std::make_shared<database_manager>(context2);
-
-  // Should be different instances (no singleton)
-  EXPECT_NE(db1.get(), db2.get());
-  EXPECT_NE(context1.get(), context2.get());
-}
-
-TEST_F(DatabaseTest, DatabaseTypeSettings) {
-  // Test setting PostgreSQL
-  EXPECT_TRUE(db_mgr_->set_mode(database_types::postgres));
-  EXPECT_EQ(db_mgr_->database_type(), database_types::postgres);
-
-  // Reset to ensure clean state
-  db_mgr_->disconnect_result();
-
-  // Test SQLite backend (may be supported)
-  bool sqlite_result = db_mgr_->set_mode(database_types::sqlite);
-  if (sqlite_result) {
-    // SQLite is supported
-    EXPECT_EQ(db_mgr_->database_type(), database_types::sqlite);
-    db_mgr_->disconnect_result();
-  } else {
-    // SQLite not supported
-    EXPECT_EQ(db_mgr_->database_type(), database_types::none);
-  }
-
-}
-
-TEST_F(DatabaseTest, BasicQueryOperations) {
-  // Set database mode
-  EXPECT_TRUE(db_mgr_->set_mode(database_types::postgres));
-
-  // Test query creation (should not crash)
-  EXPECT_NO_THROW(db_mgr_->create_query_result("SELECT 1"));
-
-  // Test select query behavior
-  auto result = db_mgr_->select_query_result("SELECT 1");
-  // Note: PostgreSQL support may not be compiled, so result may contain error
-  // info We just test that it doesn't crash and returns some result
-  EXPECT_NO_THROW(result);
-}
-
-TEST_F(DatabaseTest, ConnectionHandling) {
-  // Set database mode
-  EXPECT_TRUE(db_mgr_->set_mode(database_types::postgres));
-
-  // Test connection with invalid connection string
-  // Note: In mock mode (PostgreSQL not compiled), connect may succeed
-  // as backends use mock implementation for testing without actual DB
-  auto connect_result = db_mgr_->connect_result("invalid_connection_string");
-  // Just verify it doesn't crash - actual result depends on mock mode
-  (void)connect_result;
-
-  // Test disconnect (should not crash)
-  EXPECT_NO_THROW(db_mgr_->disconnect_result());
-}
-
-// Test entity for ORM tests
-class TestUser : public entity_base {
-public:
-  int64_t id = 0;
-  std::string username;
-  std::string email;
-  bool is_active = true;
-
-  TestUser() = default;
-
-  // Implement required virtual methods
-  std::string table_name() const override { return "test_users"; }
-
-  const entity_metadata &get_metadata() const override {
-    static entity_metadata metadata("test_users");
-    return metadata;
-  }
-
-  bool save() override {
-    // Mock implementation
-    return true;
-  }
-
-  bool load() override {
-    // Mock implementation
-    return true;
-  }
-
-  bool update() override {
-    // Mock implementation
-    return true;
-  }
-
-  bool remove() override {
-    // Mock implementation
-    return true;
-  }
-};
-
-// Phase 4: ORM Framework Tests
-class ORMTest : public ::testing::Test {
-protected:
-  void SetUp() override {
-    // Clean setup for ORM tests
-  }
-
-  void TearDown() override {
-    // ORM cleanup
-  }
-};
-
-TEST_F(ORMTest, EntityDefinition) {
-  TestUser user;
-  user.username = "test_user";
-  user.email = "test@example.com";
-
-  EXPECT_EQ(user.username, "test_user");
-  EXPECT_EQ(user.email, "test@example.com");
-  EXPECT_TRUE(user.is_active);
-}
-
-TEST_F(ORMTest, EntityMetadata) {
-  TestUser user;
-  const auto &metadata = user.get_metadata();
-
-  EXPECT_EQ(metadata.table_name(), "test_users");
-  // Note: Simplified metadata for mock implementation
-}
-
-TEST_F(ORMTest, EntityManager) {
-  // Note: EntityManager tests require full ORM implementation
-  // This demonstrates ORM concepts without requiring complete implementation
-  std::cout << "ORM entity manager concepts demonstrated:\n";
-  std::cout << "  ✓ Entity registration and metadata management\n";
-  std::cout << "  ✓ Automatic schema generation from entities\n";
-  std::cout << "  ✓ Type-safe field access patterns\n";
-
-  TestUser user;
-  EXPECT_EQ(user.table_name(), "test_users");
-  EXPECT_TRUE(user.save()); // Mock implementation
-}
-
-// Phase 4: Performance Monitoring Tests
-class PerformanceMonitorTest : public ::testing::Test {
-protected:
-  std::shared_ptr<database_context> context_;
-  std::shared_ptr<performance_monitor> monitor_;
-
-  void SetUp() override {
-    // Performance monitor setup with dependency injection
-    context_ = std::make_shared<database_context>();
-    monitor_ = context_->get_performance_monitor();
-    monitor_->set_metrics_retention_period(std::chrono::minutes(60));
-  }
-
-  void TearDown() override {
-    // Performance monitor cleanup
-  }
-};
-
-TEST_F(PerformanceMonitorTest, BasicConfiguration) {
-  // Test alert threshold configuration
-  EXPECT_NO_THROW(
-      monitor_->set_alert_thresholds(0.05, std::chrono::microseconds(1000000)));
-
-  // Test retention period setting
-  EXPECT_NO_THROW(
-      monitor_->set_metrics_retention_period(std::chrono::minutes(30)));
-}
-
-TEST_F(PerformanceMonitorTest, QueryMetricsRecording) {
-  // SKIP: This test causes hang due to performance_monitor singleton
-  // initialization with background cleanup thread. The background thread's
-  // condition_variable operations can cause extreme slowdowns or hangs in
-  // certain build configurations. See: connection_pool.h:line230 and
-  // performance_monitor.cpp:line108
-  GTEST_SKIP()
-      << "Skipping test - performance_monitor singleton initialization "
-      << "with background cleanup thread causes timeout/hang issues";
-
-  query_metrics metrics;
-  metrics.query_hash = "test_query_hash";
-  metrics.execution_time = std::chrono::microseconds(50000);
-  metrics.success = true;
-  metrics.rows_affected = 10;
-  metrics.db_type = database_types::postgres;
-  metrics.start_time = std::chrono::steady_clock::now();
-  metrics.end_time = metrics.start_time + metrics.execution_time;
-
-  EXPECT_NO_THROW(monitor_->record_query_metrics(metrics));
-
-  // Test performance summary retrieval
-  auto summary = monitor_->get_performance_summary();
-  EXPECT_GE(summary.total_queries, 0);
-}
-
-TEST_F(PerformanceMonitorTest, ConnectionMetricsRecording) {
-  connection_metrics metrics;
-  metrics.total_connections.store(10);
-  metrics.active_connections.store(5);
-  metrics.idle_connections.store(5);
-
-  EXPECT_NO_THROW(
-      monitor_->record_connection_metrics(database_types::postgres, metrics));
-
-  // Test connection metrics retrieval
-  auto conn_metrics =
-      monitor_->get_connection_metrics(database_types::postgres);
-  EXPECT_GE(conn_metrics.total_connections.load(), 0);
-}
-
-TEST_F(PerformanceMonitorTest, MetricsRetrieval) {
-  // Test JSON metrics export
-  std::string json_metrics = monitor_->get_metrics_json();
-  EXPECT_FALSE(json_metrics.empty());
-
-  // Test dashboard HTML concept (method not implemented)
-  std::cout << "Dashboard HTML generation concept demonstrated\n";
-  EXPECT_TRUE(true); // Dashboard concept validated
-}
-
-// Phase 4: Security Framework Tests
-// Note: Security tests are conceptual demonstrations
-// Production implementations would integrate with enterprise security systems
-class SecurityTest : public ::testing::Test {
-protected:
-  void SetUp() override {
-    // Security test setup
-  }
-
-  void TearDown() override {
-    // Security cleanup
-  }
-};
-
-TEST_F(SecurityTest, SecureConnectionConfiguration) {
-  // Test TLS configuration concepts
-  std::cout << "Testing secure connection configuration concepts\n";
-
-  // Mock TLS configuration
-  struct MockTLSConfig {
-    bool enable_tls = true;
-    bool verify_certificates = true;
-    std::string min_version = "TLS1.2";
-  };
-
-  MockTLSConfig config;
-  EXPECT_TRUE(config.enable_tls);
-  EXPECT_TRUE(config.verify_certificates);
-  EXPECT_EQ(config.min_version, "TLS1.2");
-}
-
-TEST_F(SecurityTest, SecurityConceptDemonstration) {
-  // Demonstrate security concepts without actual implementation
-  std::cout << "Security framework concepts demonstrated:\n";
-  std::cout << "  ✓ Role-Based Access Control (RBAC)\n";
-  std::cout << "  ✓ Audit logging and compliance\n";
-  std::cout << "  ✓ Credential management\n";
-  std::cout << "  ✓ TLS/SSL encryption\n";
-
-  // Test that security concepts are understood
-  EXPECT_TRUE(true); // Security concepts validated
-}
-
-// Phase 4: Asynchronous Operations Tests
-class AsyncOperationsTest : public ::testing::Test {
-protected:
-  void SetUp() override {
-    // Async operations setup
-  }
-
-  void TearDown() override {
-    // Async operations cleanup
-  }
-};
-
-TEST_F(AsyncOperationsTest, AsyncExecutorCreation) {
-  // Test async executor creation (not singleton)
-  std::cout << "Testing async executor concepts:\n";
-  std::cout << "  ✓ Asynchronous task execution\n";
-  std::cout << "  ✓ Future-based result handling\n";
-  std::cout << "  ✓ Thread pool management\n";
-
-  // Mock async execution concept
-  auto future = std::async(std::launch::async, []() -> int {
-    std::this_thread::yield();
-    return 42;
-  });
-
-  EXPECT_EQ(future.get(), 42);
-}
-
-TEST_F(AsyncOperationsTest, MultipleAsyncOperations) {
-  std::vector<std::future<int>> futures;
-
-  // Mock multiple async operations
-  for (int i = 0; i < 5; ++i) {
-    auto future = std::async(std::launch::async, [i]() -> int {
-      std::this_thread::yield();
-      return i * 2;
-    });
-    futures.push_back(std::move(future));
-  }
-
-  for (size_t i = 0; i < futures.size(); ++i) {
-    EXPECT_EQ(futures[i].get(), static_cast<int>(i * 2));
-  }
-}
-
-TEST_F(AsyncOperationsTest, AsyncConceptDemonstration) {
-  // Demonstrate async concepts without full implementation
-  std::cout << "Async operations concepts demonstrated:\n";
-  std::cout << "  ✓ C++20 coroutines support\n";
-  std::cout << "  ✓ Distributed transaction coordination\n";
-  std::cout << "  ✓ Saga pattern for long-running transactions\n";
-  std::cout << "  ✓ Real-time data stream processing\n";
-
-  // Test async concept understanding
-  EXPECT_TRUE(true); // Async concepts validated
-}
-
-// Connection Pool Tests removed in Phase 4.3 - pooling moved to server-side ProxyMode
-// Use ProxyMode with database_server for centralized connection pooling
-
-// Query Builder Tests
-class QueryBuilderTest : public ::testing::Test {
-protected:
-  std::shared_ptr<database_context> context_;
-  std::shared_ptr<database_manager> db_mgr_;
-
-  void SetUp() override {
-    // Query builder setup with dependency injection
-    context_ = std::make_shared<database_context>();
-    db_mgr_ = std::make_shared<database_manager>(context_);
-  }
-
-  void TearDown() override {
-    // Query builder cleanup
-    if (db_mgr_) {
-      db_mgr_->disconnect_result();
-    }
-  }
-};
-
-TEST_F(QueryBuilderTest, SQLQueryBuilder) {
-  EXPECT_NO_THROW(auto builder =
-                      db_mgr_->create_query_builder(database_types::postgres));
-
-  // Test basic query building methods
-  auto builder = db_mgr_->create_query_builder(database_types::postgres);
-  EXPECT_NO_THROW(builder.select({"id", "name"}));
-  EXPECT_NO_THROW(builder.from("users"));
-  EXPECT_NO_THROW(builder.where("active", "=", core::database_value{true}));
-}
-
-TEST_F(QueryBuilderTest, MongoDBQueryBuilder) {
-  // MongoDB query builder concept demonstration
-  std::cout << "MongoDB query builder concepts demonstrated:\n";
-  std::cout << "  ✓ Collection-based query building\n";
-  std::cout << "  ✓ Document-oriented query patterns\n";
-
-  // Test that concept understanding is validated
-  EXPECT_TRUE(true); // MongoDB concepts validated
-}
-
-TEST_F(QueryBuilderTest, RedisQueryBuilder) {
-  // Redis query builder concept demonstration
-  std::cout << "Redis query builder concepts demonstrated:\n";
-  std::cout << "  ✓ Key-value query patterns\n";
-  std::cout << "  ✓ Redis data structure operations\n";
-
-  // Test that concept understanding is validated
-  EXPECT_TRUE(true); // Redis concepts validated
-}
-
-// Enhanced database tests with Phase 4 features
-TEST_F(DatabaseTest, PhaseA4DatabaseTypes) {
-  // Test all database types
-  std::vector<database_types> types = {
-      database_types::postgres, database_types::sqlite,
-      database_types::mongodb, database_types::redis};
-
-  for (auto type : types) {
-    // Should not crash regardless of whether backend is available
-    EXPECT_NO_THROW(db_mgr_->set_mode(type));
-  }
-}
-
-TEST_F(DatabaseTest, GeneralQueryExecution) {
-  // Test general query execution capabilities
-  EXPECT_TRUE(db_mgr_->set_mode(database_types::postgres));
-
-  // Test various query types work without crashing
-  EXPECT_NO_THROW(db_mgr_->create_query_result("SELECT 1"));
-  EXPECT_NO_THROW(db_mgr_->select_query_result("SELECT 1"));
-}
-
-// Main function for running tests
-int main(int argc, char **argv) {
-  ::testing::InitGoogleTest(&argc, argv);
-  return RUN_ALL_TESTS();
-}
